Re-organize attachment factories
