+++ UNRELEASED_CHANGELOG.md
@@ -86,6 +86,9 @@
 
 ### ⚠️ Changed
 - `StreamAttachment.defaultFactories()` is a function now, instead of a property.
+- Updated all default value factories to functions (e.g. StreamTypography)
+- Re-organized all attachment factories and split up code in multiple packages
+-
 
 ### ❌ Removed
