+++ stream-chat-android-compose/src/main/java/io/getstream/chat/android/compose/ui/attachments/link/LinkAttachmentFactory.kt
@@ -1,9 +1,8 @@
-package io.getstream.chat.android.compose.ui.attachments.factory
+package io.getstream.chat.android.compose.ui.attachments.link
 
 import androidx.compose.runtime.Composable
 import com.getstream.sdk.chat.model.ModelType
 import io.getstream.chat.android.compose.ui.attachments.AttachmentFactory
-import io.getstream.chat.android.compose.ui.attachments.content.LinkAttachmentContent
 import io.getstream.chat.android.compose.ui.util.hasLink
 
 /**
